# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Sociomile Support API
  version: 1.0.0
  description: Multi-tenant Customer Support Backend API

servers:
  - url: http://localhost:8080

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  parameters:
    TenantHeader:
      in: header
      name: X-Tenant-ID
      required: true
      schema:
        type: string

  schemas:

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        token:
          type: string

    WebhookRequest:
      type: object
      required: [external_customer_id, message_body]
      properties:
        external_customer_id:
          type: string
        customer_name:
          type: string
        message_body:
          type: string

    Conversation:
      type: object
      properties:
        id: { type: integer }
        tenant_id: { type: integer }
        customer_id: { type: integer }
        assigned_agent_id: { type: integer, nullable: true }
        status: { type: string }
        last_message_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }

    Message:
      type: object
      properties:
        id: { type: integer }
        conversation_id: { type: integer }
        sender_type: { type: string }
        body: { type: string }
        created_at: { type: string, format: date-time }

security:
  - bearerAuth: []

paths:

  /api/auth/login:
    post:
      summary: Login
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: OK

  /api/webhooks/channel:
    post:
      summary: Receive channel webhook
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
      responses:
        201:
          description: Conversation created or reused

  /api/conversations:
    get:
      summary: List conversations
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: query
          name: status
          schema: { type: string }
      responses:
        200:
          description: OK

  /api/conversations/{id}:
    get:
      summary: Get conversation detail
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: OK

  /api/conversations/{id}/assign:
    patch:
      summary: Assign agent
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: integer
      responses:
        200:
          description: Assigned

  /api/conversations/{id}/status:
    patch:
      summary: Update status
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string

  /api/conversations/{id}/messages:
    get:
      summary: List messages
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: path
          name: id
          required: true
          schema: { type: integer }

    post:
      summary: Send agent message
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string

  /api/customers:
    get:
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/TenantHeader'

    post:
      summary: Create customer
      parameters:
        - $ref: '#/components/parameters/TenantHeader'

  /api/customers/{id}:
    get:
      summary: Get customer
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: path
          name: id
          required: true
          schema: { type: integer }